+++ src/dsp/windowing.cpp
@@ -18,12 +18,12 @@ dataBuf applyBlackmanHarris(dataBuf buf, float a0, float a1, float a2, float a3)
     return newBuf;
 }
 
-float applyBlackmanHarrisToSingleValue(float value, int bufferSize, int bufferIndex)
+double applyBlackmanHarrisToSingleValue(double value, int bufferSize, int bufferIndex)
 {
-    float val = 0.35875;
-    val -= 0.48829 * cos( (2.0 * M_PI * bufferIndex) / (float)bufferSize );
-    val += 0.14128 * cos( (4.0 * M_PI * bufferIndex) / (float)bufferSize );
-    val -= 0.01168 * cos( (6.0 * M_PI * bufferIndex) / (float)bufferSize );
+    double val = 0.35875;
+    val -= 0.48829 * cos( (2.0 * M_PI * bufferIndex) / ((float)bufferSize) );
+    val += 0.14128 * cos( (4.0 * M_PI * bufferIndex) / ((float)bufferSize) );
+    val -= 0.01168 * cos( (6.0 * M_PI * bufferIndex) / ((float)bufferSize) );
 
     return value * val;
 }
