+++ src/include/dsp/fft.hpp
@@ -5,4 +5,4 @@
 #include <math.h>
 #include <complex>
 
-std::shared_ptr<std::vector<std::complex<float>>> fft(std::shared_ptr<std::vector<std::complex<float>>> dataPoints, float fftSize);
+std::shared_ptr<std::vector<std::complex<double>>> fft(std::shared_ptr<std::vector<std::complex<double>>> dataPoints, float fftSize);
